Implements all API methods in ScreenMain and ScreenList, renames amount to value and type to name throughout the system

+++ client/components/ScreenList/ScreenList.jsx
@@ -4,7 +4,8 @@ import Layout from '../Layout/Layout.jsx'
 import SpentForm from '../SpentForm/SpentForm.jsx'
 import SpentList from '../SpentList/SpentList.jsx'
 import Title from '../Title/Title.jsx'
-import listsStore from '../../modules/lists-store'
+import api from '../../modules/api'
+import cloneDeep from 'lodash/cloneDeep'
 
 export default class ScreenList extends Component {
   constructor (props) {
@@ -12,7 +13,8 @@ export default class ScreenList extends Component {
 
     this.state = {
       readyToDeleteId: null,
-      listId: Number(this.props.match.params.id)
+      listId: Number(this.props.match.params.id),
+      isLoaded: false
     }
 
     this.addItem = this.addItem.bind(this)
@@ -26,19 +28,37 @@ export default class ScreenList extends Component {
   }
 
   updateState () {
-    this.setState({
-      currentList: listsStore.listGet(this.state.listId)
-    })
+    api.getList(this.state.listId)
+      .then(({data}) => {
+        this.setState({
+          currentList: data,
+          isLoaded: true
+        })
+      })
   }
 
   addItem (item) {
-    listsStore.itemAdd(this.state.listId, item)
-    this.updateState()
+    api.createItem(this.state.listId, item)
+      .then(({data}) => {
+        this.setState(prevState => {
+          let newState = cloneDeep(prevState)
+          newState.currentList.items.unshift(data)
+          newState.currentList.itemsCount++
+          return newState
+        })
+      })
   }
 
   deleteItem (id) {
-    listsStore.itemDelete(this.state.listId, id)
-    this.updateState()
+    api.deleteItem(this.state.listId, id)
+      .then(({data}) => {
+        this.setState(prevState => {
+          let newState = cloneDeep(prevState)
+          newState.currentList.items = newState.currentList.items.filter(({id}) => id !== data.itemId)
+          newState.currentList.itemsCount--
+          return newState
+        })
+      })
   }
 
   setReadyToDelete (id) {
@@ -50,26 +70,30 @@ export default class ScreenList extends Component {
   }
 
   getContent () {
-    const items = this.state.currentList.items
+    if (this.state.isLoaded) {
+      const items = this.state.currentList.items
 
-    if (items.length) {
-      return (
-        <SpentList
-          items={items}
-          readyToDeleteId={this.state.readyToDeleteId}
-          onReadyToDelete={this.setReadyToDelete}
-          onItemDelete={this.deleteItem} />
-      )
+      if (items.length) {
+        return (
+          <SpentList
+            items={items}
+            readyToDeleteId={this.state.readyToDeleteId}
+            onReadyToDelete={this.setReadyToDelete}
+            onItemDelete={this.deleteItem} />
+        )
+      }
     }
   }
 
   getTitle () {
-    return (
-      <Title
-        back={<Link to='/'>To the main screen</Link>}>
-        {this.state.currentList.name}
-      </Title>
-    )
+    if (this.state.isLoaded) {
+      return (
+        <Title
+          back={<Link to='/'>To the main screen</Link>}>
+          {this.state.currentList.name}
+        </Title>
+      )
+    }
   }
 
   render () {
